requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
supabase>=2.3.0
folium>=0.16.0
streamlit-folium>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
